End of day implementation
I created a timer for ending all the sales, then it saves all the data
to a new xml called "endOfDay.xml". theoretically we'll need to take
that data the next time the app runs.
Right now there is a problem when saving the data. it only saves 5 sales
even though the size of the list array is 10! maybe something to do with
synchronizing.